Day 7 solution
Went to the subreddit for part 1. Not for answers, but for additional
test input while I figured out what I was doing wrong on test input
parsing. Turns out I forgot to test all three array entries for the
three group case...

For part 2 I just added a generic use an alternate ranking method that
added J to the largest group count, and transformed J -> 1 in the hand
to re-use the existing comparison.

Lots of attempts to self-implement Ord and PartialOrd before I got a
hang of using the Rust defaults.