Day 17 Part 2 Solution
Part 2 was actually very solvable with Z3. The original int-based
analysis took too long to run, but all operations could be described as
bit operations, so it ran to a pretty snappy conclusion on bit vector
inputs and constraints only.

This marks the first time I've used Z3. Ownership is a bit clunky in the
Rust struct model (at least as far as I've constructed it), but this
library is pretty useful.